requirements-dev.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
black==23.3.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
flake8==6.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.39
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.18.4
ruamel.yaml.clib==0.2.8
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.9
